+++ autosub/__init__.py
@@ -128,15 +128,6 @@ def main():  # pylint: disable=too-many-branches, too-many-statements, too-many-
                         print(_("Audio pre-processing complete."))
 
             else:
-                args.audio_split_cmd = \
-                    args.audio_split_cmd.replace(
-                        "[channel]",
-                        "{channel}".format(channel=args.api_audio_channel))
-                args.audio_split_cmd = \
-                    args.audio_split_cmd.replace(
-                        "[sample_rate]",
-                        "{sample_rate}".format(sample_rate=args.api_sample_rate))
-
                 if args.audio_split_cmd == constants.DEFAULT_AUDIO_SPLT:
                     # if user doesn't modify the audio_split_cmd
                     if args.api_suffix == ".ogg":
@@ -151,6 +142,15 @@ def main():  # pylint: disable=too-many-branches, too-many-statements, too-many-
                             args.audio_split_cmd.replace(
                                 "-vn",
                                 "-vn -c:a pcm_s16le -f s16le")
+
+                args.audio_split_cmd = \
+                    args.audio_split_cmd.replace(
+                        "[channel]",
+                        "{channel}".format(channel=args.api_audio_channel))
+                args.audio_split_cmd = \
+                    args.audio_split_cmd.replace(
+                        "[sample_rate]",
+                        "{sample_rate}".format(sample_rate=args.api_sample_rate))
 
             cmdline_utils.validate_aovp_args(args)
             fps = cmdline_utils.get_fps(args=args, input_m=input_m)